requirements.txt
aiosqlite~=0.20.0
cachetools~=5.3.3
python-dotenv~=1.0.1
python-telegram-bot>=21.0
httpx>=0.27.0
websockets>=12.0
requests>=2.31.0
